requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
mlflow
azureml-core
azureml-defaults
